Adds find_element Function
This is a semi-revined version of something that I worked on a few years ago. It takes in a dumped JSON structure and allows you to specify a path through it to find particular elements by passing in a list of the items you are searching for. You can also get all elements at a level (have it return an array for that level) by specifying the symbol "*". From there, you can have the function return strings, dictionaries, or other valid JSON elements and then operate on them with native Python from there.